+++ src/ui/buttons/props.js
@@ -1,4 +1,4 @@
-/* eslint-disable max-lines, eslint-comments/disable-enable-pair  */
+/* eslint-disable eslint-comments/disable-enable-pair  */
 /* @flow */
 
 import { ZalgoPromise } from '@krakenjs/zalgo-promise/src';
@@ -393,7 +393,8 @@ export type RenderButtonProps = {|
     applePaySupport : boolean,
     supportsPopups : boolean,
     supportedNativeBrowser : boolean,
-    experience : string
+    experience : string,
+    showPayLabel : boolean
 |};
 
 export type PrerenderDetails = {|
@@ -489,7 +490,8 @@ export type ButtonPropsInputs = {
     applePaySupport : boolean,
     supportsPopups : boolean,
     supportedNativeBrowser : boolean,
-    experience : string
+    experience : string,
+    showPayLabel : boolean
 };
 
 export const DEFAULT_STYLE = {
@@ -659,7 +661,8 @@ export function normalizeButtonProps(props : ?ButtonPropsInputs) : RenderButtonP
         applePaySupport = false,
         supportsPopups = false,
         supportedNativeBrowser = false,
-        experience = ''
+        experience = '',
+        showPayLabel = true
     } = props;
 
     const { country, lang } = locale;
@@ -702,5 +705,5 @@ export function normalizeButtonProps(props : ?ButtonPropsInputs) : RenderButtonP
 
     return { clientID, fundingSource, style, locale, remembered, env, fundingEligibility, platform, clientAccessToken,
         buttonSessionID, commit, sessionID, nonce, components, onShippingChange, onShippingAddressChange, onShippingOptionsChange, personalization, content, wallet, flow,
-        experiment, vault, userIDToken, applePay, applePaySupport, supportsPopups, supportedNativeBrowser, experience };
+        experiment, vault, userIDToken, applePay, applePaySupport, supportsPopups, supportedNativeBrowser, experience, showPayLabel };
 }